Adding raw_socket callback to cowboy_req
This allows to drop into raw mode with cowboy without exiting the
handler in order to drop to whatever transport underlies the connection.

Users of this mode must be sure they do not want to reuse cowboy to
parse headers, responses, or whatever, and return a clean connection (or
a closed one) when returning the handler.