Ignore some never-working auto-responses.
Because of some of the code paths we use in vegur to send content by
delaying responses that we may send manually without always respecting
the internal cowboy paths (i.e. hijacking the connection for websocket
pipelines), it is possible that the impacted path will have sent data
over the line, without having gone through the regular self-messages
that cowboy uses.

Alternatively, other paths will actually rely on that behaviour working
for specific failures.

This patch makes it so that the path does detection of useless replies
that would otherwise crash and generate errors, and silences them
instead.